Angefangen einen WAV-File-Player zu bauen (noch nicht fertig!)

+++ Wav-File IO/src/wenzel/paul/speechsynthesis/model/dataobjects/WavFileDataObject.java
@@ -88,7 +88,7 @@ public class WavFileDataObject {
 	
 //////////////////////////////////////////////////Methoden///////////////////////////////////////////////////
 	
-	public long getNumberOfFrames() {
+	public int getNumberOfFrames() {
 		return wavFileValues[0].length;
 	}
